Move getOrderDescription to the routes directory
AB#8867
